refactored sig verif
